we ensure that the `stored.rules`-file is sorted alphabetically
In the file `stored.rules` there are all frozen rules stored, along with an uuid to a file where than all frozen violations for that rule are stored.
Currently, the content of that file is not sorted in any way.
Therefore, adding a new frozen rule adds a new line to that file and can lead to a reorder of all lines.
The reorder of the file is especially than a problem, when we want to see the changed to the previous version as a diff.
Ideally by adding one new rule also only one line should change in the file.
By sorting the file alphabetically this is given.

Used the following post to sort the file: https://stackoverflow.com/a/55957344

Signed-off-by: Hannes Oberprantacher <h.oberprantacher@gmail.com>