deprecate `ConditionEvents` methods that impede maintenance
Originally `ConditionEvents` were designed quite generically to be used for simple rules, but also within `Contains{Any/Only}ElementCondition` or for negation. For this we needed methods like `getAllowed()` or `isEmpty()` to invert the condition or check if any element violated the condition. However, the implementation is quite suboptimal (in many cases the allowed events are unnecessary but clutter memory because the number of allowed events is most of the time magnitudes bigger than the number of violations). This is likely a public API hardly any user really needs (we will see if anybody files an issue) but that prevents us from improving the memory footprint when checking conditions.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>