make `LayeredArchitecture` immutable
For all properties except for `optionalLayers` we have followed the pattern of immutability, we should do the same here. Technically this is a breaking change, but since this immutability is the usual pattern that ArchUnit follows everywhere else, and the fluent API invites to simply chain these methods anyway, the risk is probably low.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>