remove configuration `archRule.failOnEmptyShould=false` for tests
It turned out to be a bad idea to override the default configuration for `archRule.failOnEmptyShould=false` for all tests, since that covered up the fact that the default configuration actually breaks `optionalLayers` for `LayeredArchitecture` and `OnionArchitecture`. We now also use the default configuration for tests and override it on an individual basis. Note that we should use the default config for `ArchitecturesTest`, but at the moment the tests would be broken. So we deactivate it for now and fix it together with making `failOnEmptyShould` configurable on a per-rule basis.

Note that this also showed some tests that were a little shady, i.e. tests that checked for `assertThatMembers(..).matchInAnyOrderMembersOf(..)`, but actually those members were an empty set, at least with certain JDK versions. For members declared in meta-annotated types, this actually also hits the annotation type annotated with the meta-annotation, because meta-annotated counts direct annotations as well. Since the behavior always was like this we adjust the test to match reality in this case.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>