make `Members-/Methods-/...That` reusable in other contexts
By specifying bounds like `MethodsThat<CONJUNCTION extends GivenMethodsConjunction>` on the type `MethodsThat` we are tying this to the `methods()` API. This way we cannot use the syntax element in another context, e.g. `OnlyBeCalledBy -> MethodsThat`. We should control the bound on the methods within the `members()/methods()/...` API to keep the elements themselves reusable, but still support all the specific syntax paths for the subtypes (e.g. `methods.that().areFinal()` which would not work on `members().that()`).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>