fix detecting multiple identical accesses in same line
Two accesses should not be equal if origin, target and line number match. Because it is legal to write something like

```
void call() {
  callSomeMethod.chainSameMethod().chainSameMethod().chainSameMethod();
}
```

In such a case origin, target and line number are identical. Thus, when we return a `Set<JavaCall<?>>` we will only contain one element with the old implementation, which is not correct. Since only ArchUnit's importer creates domain objects we can simply use the identity `equals` and `hashCode`. Because in the end each access is unique by itself and users never create these objects themselves to compare them against imported objects.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>