add JUnit5-support for `@ArchTest` members in abstract base classes
So far, having `@ArchTest` fields or methods in an abstract base class would cause the JUnit 5 engine to crash. This also made the pattern to "share" tests via extending a common base class impossible, which some users would like to have.

The change corresponds to the commits
* 8188cdc68abd9bbedd87a9800d89e6a260e3a52d
* 625c3d6fe42781a5f62c4d4ffe103d1376b338e4
* 59f8c4da95fb95a4be49e9aad2311b3af06edd26
which added the same feature for JUnit4 (Issue: #104).

Signed-off-by: Manfred Hanke <Manfred.Hanke@tngtech.com>