replace `Iterable` in user APIs by `Collection`
Since the API to create a `Stream` from an `Iterable` is so much less user-friendly we should try to offer `Collection` in the API where reasonable. APIs that consume from the user should take `Iterable` if there is no reason for a more narrow type. But APIs that offer to the user should offer `Collection` instead of `Iterable` to make the `Stream` processing simpler.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>