only import necessary classes when discovering tests from classpath
The JUnit platform API makes it necessary for ArchUnit to provide an implementation to discover test classes from the classpath (e.g. because the discovery request contains a package selector). Since ArchUnit itself is perfectly suited for this and does not demand any third party dependency we simply use the `ClassFileImporter` for this. However, so far we have used the current `ArchConfiguration` of the user, which by default e.g. resolves all further dependencies from the classpath. This is unnecessary overhead and can substantially slow down the discovery process if the test classes e.g. have some framework dependencies like the Spring framework.
We will now switch the `ArchConfiguration` for the discovery process. However, the global nature of `ArchConfiguration` does not make this completely trivial. E.g. if the user has another import in progress in another thread we do not want to affect this import by suddenly switching the user's configuration in the middle. To isolate this configuration change I have added a possibility to override the `ArchConfiguration` thread locally. I.e. the discovery process will only affect the current thread when changing the configuration and only during the scope of a fixed lambda. Afterwards the thread local configuration is cleaned up. This way we should never affect any other import of a user, since the scope of configuration change is tailored exactly to the discovery process and cleaned up afterwards.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>