make `ViolationHandler` API safe for use with lambdas
For a Java 7 language level the API `handleViolations(ViolationHandler<T> handler)` was quite fitting, since there was a good chance that this would be an anonymous instantiation capturing the reified type `T`. However, with Java 8 language level the most straight forward API would accept a lambda, and in that case we would typically lose that information.
We will add a hack to ensure that the type parameter of `ViolationHandler<T>` will always be reified by adding an unused varargs array of that type to the method. While this is not very pretty, it is the only way I found to really force the type parameter to be reified and not get erased. Since the usage of the API is then quite convenient (e.g. it is possible to just write `handleViolations((Collection<JavaAccess<?>> items, String message) -> ...)` and the items will automatically be filtered for `JavaAccess`) I decided to still go the way of this "hack" with some proper documentation.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>