make `Formatters.formatNamesOf(..)` compatible with bound types
At the moment if there is some sort of `Collection<Class<? extends SomeBound>>` instead of `Collection<Class<?>>` then `Formatters.formatNamesOf(collection)` does not compile, because `Class<? extends SomeBound>` counts as subtype of `Class<?>`. Thus, we change the signature to allow `? extends Class<?>` to make this work out of the box with bounded types.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>