improve javadoc on `MembersThat.areDeclaredInClassesThat(..)`
The example was a little confusing since it looks like `areSubtypeOf` would be a predicate from the ArchUnit standard library, while it in fact was meant as a custom predicate. I have now replaced this by a predicate that really exists and made the example a little bit more realistic.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>