add support for private accesses to other classes
For private accesses like

```
class Outer {
  // ...
  void callInner() {
    inner.call();
  }

  class Inner {
    private void call() {
    }
  }
}
```

the compiler would create synthetic access methods instead of adding a direct access. I.e. in the upper case we would see something like `Outer -> Inner.access$000() -> Inner.call()`. This is confusing to users as it does not mirror the source code. Thus, similar to lambdas we will fix this by merging the call to the synthetic access method and the access from the synthetic access method into a single access with correct origin and target.

Note that for the moment we keep the synthetic constructor call for this scenario. I.e. if a class calls a private constructor the compiler adds a synthetic accessible constructor with some synthetic marker parameter. But for now we will not touch this, because a) the fix would be quite complex in comparison to `access$xxx` handing, b) it is less confusing than the `access$xxx` method, since it is almost the same constructor call except for one additional parameter and c) we do not know if users ever care about this anyway, so we should wait and see if this ever causes any real problem.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>