allow `ClassResolverFactory` to instantiate private `ClassResolvers`
So far `ClassResolver` implementations had to have a constructor that is accessible to the `ClassResolverFactory`. This can be a little limiting if we just want to reconfigure the `ClassResolver` with a private class in a narrow scope. Thus, we now use Reflection to allow access also to non-accessible (e.g. private) classes.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>