add `ArchCondition` to check for transitive dependencies
The original implementation had the problem that it reported every single transitive dependency that matched the predicate. This would mean, even if a transitive dependency got reported as violation, all further transitive dependencies of that dependency would still be analyzed and potentially also reported as violation. For rules like

```
noClasses()
  .that().resideInAPackage("..example..")
  .should().transitivelyDependOnClassesThat().resideInAPackage("..some.framework..")
```

this would mean thousands of violations, even for classes that are only reached through other framework classes.

We now do a depth first search approach for all dependency paths that are outgoing from the analyzed classes until we find a matching target. Then we stop and report the path. This way all further dependencies will be omitted and the analysis will complete a lot faster.

Issue: #780
Signed-off-by: e.solutions <17569373+Pfoerd@users.noreply.github.com>
on-behalf-of: @e-esolutions-GmbH <info@esolutions.de>