split `ArchitecturesTest` into separate test classes
The tests for onion architecture and layered architecture are quite unrelated except for some common test infrastructure. In the end these are two classes housed inside the same umbrella class, but that does not mean we have to test them in one single test class. Having a single test class makes it harder to understand which test method belongs to which class.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>