fix `PlantUmlArchCondition.Configuration` not accessible warning
On the contrary to the Java compiler the Kotlin compiler emits a warning that `PlantUmlArchCondition.Configuration` as the parameter type of `PlantUmlArchCondition.adhereToPlantUmlDiagram()` is not accessible (because it is package private). Since in the end the pattern used here is a little inconsistent to other places anyway we change this now to be a publicly accessible class. The current construct was a little strange, either it needs to be a generic interface, freely extensible, but then there does not seem to be a reason to hide it from the user. Or we want to limit the possibilities, but then we might as well use a concrete class. As an additional benefit (since we will break the API anyway with release 1.0) we can now define the factory methods directly on the parameter type removing one indirection.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>