remove accesses to synthetic enum `SwitchMap` fields
For every enum switch statement the compiler generates some synthetic class with a switch-map field that contains the mapping of enum ordinals to switch case indexes (to translate an enum switch to an integer switch under the hood). This lookup is irrelevant for ArchUnit users and only creates confusion, since it is not visible in the source code. All accesses from an enum switch statement are correct already, since enum case statements are just direct parts of the method declaring the switch statements, so the origin location is correct.
We ignore these synthetic field accesses in a rather unsophisticated way (simply ignoring all field accesses to fields starting with `$SwitchMap$`), but at this point of the import we do not have enough information yet. On the other hands it is highly unlikely that users would define fields with names that start with `$SwitchMap$` as it goes against all conventions and in the end could also lead to compile errors if the name would then clash with an auto generated method (same as defining methods that clash with the synthetic `lambda$...` pattern).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>