add check for modified files to CI build
During the build we ensure some conventions, e.g. that the respective copyright header is created on new files. If developers do not run the build locally and just create a PR we need to ensure that such files cannot pass the CI build. I.e. if at the end of the normal CI build we end up with modified files there are some files checked in that do not adhere to project conventions and are thus "fixed" during the CI run.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>