add primitive classes to package `java.lang`
So far primitive types like `int` have been in the default package `""`, since their `Package` object from the Reflection API is `null` and that seemed like the best matching ArchUnit version (since ArchUnit does not use `null` within the public API and the choice would then just have been to make the `JavaPackage` `Optional` just because of a couple of corner cases). Starting with Java 9 the Reflection API introduced a new method `Class.getPackageName()` though, which would return `java.lang` for primitives and the package of the component type for arrays. Luckily for arrays ArchUnit is already consistent, because we decided long ago that this behavior for the package name would make sense. This commit now makes the package of primitives again consistent with the Reflection API by putting primitive types into the package `java.lang`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>