upgrade `actions/setup-java` from `1` to `3.6.0`
We now need to specify a distribution explicitly or the action will throw an error.
We picked `zulu`, since that was the one we have used so far anyway with the `v1`
and it offers the full range of Java versions so far.

Note that the release [v3.6.0](https://github.com/actions/setup-java/releases/tag/v3.6.0)
notably contains the 'Update `@actions/core` to 1.10.0'`, i.e.
uses the file version of `save-state` and `set-output` commands.

The stdout versions are deprecated, see
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>
Signed-off-by: Manfred Hanke <Manfred.Hanke@tngtech.com>