improve javadoc on rules API
In particular, users were confused about the semantics of joining multiple `andShould()` and `orShould()`. It is not self-evident that there is no operator precedence as developers are used to from `&&` and `||`. We should make sure that every conjunction at least links to some documentation that explains the associativity of joining multiple parts, as well as the alternative by directly supplying predicates/conditions that can be freely nested and joined via `and(..)` and `or(..)`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>