improve javadoc on predicates
At the moment it is quite challenging to find a certain `DescribedPredicate`, because users might need to know the domain model very well to find the necessary location. While, in the long run we might offer a better entry point into the API that improves discoverability of predicates, as a slight improvement we add javadoc to all predicates linking to respective other predicates that can also be applied to the object. E.g. while before users might be confused why `JavaClass.Predicates` has no predicate defined for `JavaClass.name` now there is at least a link that further predicates that can be applied to `JavaClass` due to inheritance can be found in `HasName.Predicates`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>