fix type bounds of `onlyBeCalledBy...That(..)`
The type bounds were too limiting. If we want to restrict to `onlyBeCalledByMethodsThat(predicate)`, then we should support a `DescribedPredicate<T>` up to `T` being `JavaMethod` and not just `JavaMember`. Otherwise, it is impossible to assert things like `onlyBeCalledByMethodsThat(haveRawParameterTypes(..))`, because `haveRawParameterTypes(..)` is a predicate on `JavaCodeUnit` and not `JavaMember`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>