support `-rc` suffixes as release versions
So far we assumed that any suffix would just be `-SNAPSHOT` anyway, so dropping the suffix for the release was sufficient. If we want to be able to release `1.0.0-rc1`, then we need to support a release suffix instead of dropping everything. By convention the difference between a version "in progress" and ready for release is just the `-SNAPSHOT` suffix though. We do not use any other convention. So, it makes sense to just drop any `-SNAPSHOT` suffix and assume that is our release version. That will work for all existing cases, but also for this new `-rc` case now.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>