update ArchUnit-Example sources for release test
At the moment the release test only supports non-breaking changes for the ArchUnit-Example project. While this in general is a nice idea it now restricts us to release breaking changes with `1.0.0-rc1`. I added an update of the ArchUnit-Example sources, just like the real update of the ArchUnit-Example project once the release has been published. By that we are testing against the current ArchUnit-Example sources that have been adjusted to possible breaking changes, which is i.g. fine since the release test is focused on functionality, not backwards compatibility.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>