improve naming around `JavaPackage` dependencies
The current implementation had quite some ambiguities. In particular what "the package" and the "classes within the package" refer to. Take "class dependencies from self", this could easily be understood as "the dependencies to classes that originate from *this* package", where "this package" means the classes that directly reside in the package. This would also be consistent with the Java Language Specification. To disambiguate these cases we now clearly distinguish if we are talking about dependencies from/to *this* package (i.e. classes directly within the package) and the package tree (i.e. all classes in this package or subpackages, subpackages of subpackages, ...).

Resolves: #919

Signed-off-by: Gediminas Rimša <gediminas.rimsa@gmail.com>