Make AnalyzeClasses inherited
I'd propose this change so a common configuration for the AnalyzeClasses annotation can be set in a common superclass for ArchUnit test classes within a project:
```
@AnalyzeClasses(packages = "com.example.service", importOptions = {DoNotIncludeTests.class, DoNotIncludeSomething.class})
@Target(ElementType.TYPE)
@Retention(RetentionPolicy.RUNTIME)
@Inherited
public @interface ArchRulesTest {
    ...
}

@ArchRulesTest
public class SecurityArchitecturePolicies {
    ...
}

@ArchRulesTest
public class MappingFrameworkPolicies {
    ...
}

@ArchRulesTest
public class LayeredArchitecturePolicies {
    ...
}
```
I haven't actually tested this, so this is just a recommendation - if you'd like me to fork the repo, create tests, etc. please let me know and I'll see when I'll be able to get to it.