improve Javadoc on `JavaPackage` and `HasAnnotations`
As GitHub discussions show the nature of what a package, subpackages, a package tree, etc., is actually not that self-evident. Thus, we should make sure to document the methods thoroughly and explain for each of these expressions what they mean within ArchUnit.
We will improve the API for dependencies in a later step, thus we do not need to adjust the Javadoc for those methods for now.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>